feat: [US-013] - Rename snake_case identifiers in server/
